% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sendwithus_list_templates}
\alias{sendwithus_list_templates}
\title{Get an list of templates}
\usage{
sendwithus_list_templates(api_key)
}
\arguments{
\item{api_key}{Your Stripe API Key}
}
\value{
A list of templates
}
\description{
This returns a list of templates in sendwithus
}

