% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sendwithus_send}
\alias{sendwithus_send}
\title{Send an email with template}
\usage{
sendwithus_send(api_key, template_id, variables, recipient, sender,
  esp = NULL, cc = NULL, bcc = NULL, tags = NULL, inline = NULL,
  files = NULL)
}
\arguments{
\item{api_key}{api_key for this}

\item{template_id}{The id for the template you want to send}

\item{variables}{This is a list containing variables with the variable name as the list item name}

\item{recipient}{This is a list containing the following params
\describe{
   \item{name}{The name of the person}
   \item{address}{The address of the person}
}}

\item{sender}{This is a list containing the following params
\describe{
   \item{name}{The name of the company/person sending}
   \item{address}{The address of the company/person sending}
   \item{reply_to}{The address of the company/person to reply to.}
}}

\item{esp}{The id of the ESP to send through}

\item{cc}{A vector containing cc emails}

\item{bcc}{A vector containing bcc emails}

\item{tags}{A vector containing tags}

\item{inline}{A list containing the following
\describe{
   \item{id}{The id of the file, e.g 'cat.png'},
   \item{data}{base_64_encoded_file_data}
}}

\item{files}{A data frame containing the following
\describe{
   \item{id}{The id of the file, e.g 'cat.png'},
   \item{data}{base_64_encoded_file_data}
}}
}
\value{
json with status equal to ok
}
\description{
This sends an email with template
}

